Library Management
The project is an example of a library management system.
LibraryHolding.java describes an abstract class
that represents a generic library item.  Book.java and DVD.java  extend
LibraryHolding. User.java describes the User class and Library.java
describes the Library class. A Library contains a collection of items
and a collection of users and some methods to operate on these
collections.